UPDATEL:add separate component file
