+++ apps/frontend/src/components/Category-Componanent/ProductCard.tsx
@@ -1,4 +1,4 @@
-import type { Product } from "../types/Product";
+import type { Product } from "../../types/Product";
 import React from "react";
