+++ apps/frontend/src/pages/UserDashboard-pages/UserDashboard.tsx
@@ -1,10 +1,10 @@
 import React, { useState } from "react";
-import Sidebar from "../components/Sidebar";
-import DetailsPanel from "../components/DetailsPanel";
-import OrdersPanel from "../components/OrdersPanel";
-import ServicePanel from "../components/ServicePanel";
-import type { UserProfile, ActivePanel } from "../types/dashboard.types";
-import { MOCK_USER, MOCK_ORDERS, MOCK_WISHLIST, MOCK_SERVICE_REQUESTS } from "../data/mockData";
+import Sidebar from "../../components/User_Dashboard-Component/Sidebar";
+import DetailsPanel from "../../components/User_Dashboard-Component/DetailsPanel";
+import OrdersPanel from "../../components/User_Dashboard-Component/OrdersPanel";
+import ServicePanel from "../../components/User_Dashboard-Component/ServicePanel";
+import type { UserProfile, ActivePanel } from "../../types/dashboard.types";
+import { MOCK_USER, MOCK_ORDERS, MOCK_WISHLIST, MOCK_SERVICE_REQUESTS } from "../../data/mockData";
 
 export default function UserDashboard() {
   const [activePanel, setActivePanel] = useState<ActivePanel>("details");
